fix documentation generation
