+++ inject-java/src/main/java/io/micronaut/annotation/processing/ServiceDescriptionProcessor.java
@@ -15,6 +15,22 @@
  */
 package io.micronaut.annotation.processing;
 
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.HashMap;
+import java.util.HashSet;
+import java.util.List;
+import java.util.Map;
+import java.util.Optional;
+import java.util.Set;
+
+import javax.annotation.processing.RoundEnvironment;
+import javax.annotation.processing.SupportedOptions;
+import javax.lang.model.element.Element;
+import javax.lang.model.element.TypeElement;
+import javax.lang.model.type.DeclaredType;
+import javax.lang.model.type.TypeMirror;
+
 import io.micronaut.annotation.processing.visitor.JavaClassElement;
 import io.micronaut.context.ApplicationContextConfigurer;
 import io.micronaut.context.annotation.ContextConfigurer;
@@ -23,26 +39,6 @@ import io.micronaut.core.annotation.AnnotationMetadata;
 import io.micronaut.core.annotation.AnnotationValue;
 import io.micronaut.core.annotation.Generated;
 import io.micronaut.core.util.StringUtils;
-import io.micronaut.inject.configuration.ConfigurationMetadataBuilder;
-import io.micronaut.inject.configuration.ConfigurationMetadataWriter;
-
-import javax.annotation.processing.RoundEnvironment;
-import javax.annotation.processing.SupportedOptions;
-import javax.lang.model.element.Element;
-import javax.lang.model.element.TypeElement;
-import javax.lang.model.type.DeclaredType;
-import javax.lang.model.type.TypeMirror;
-import java.io.IOException;
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.HashMap;
-import java.util.HashSet;
-import java.util.List;
-import java.util.Map;
-import java.util.Optional;
-import java.util.ServiceConfigurationError;
-import java.util.ServiceLoader;
-import java.util.Set;
 
 /**
  * A separate aggregating annotation processor responsible for creating META-INF/services entries.
@@ -94,17 +90,18 @@ public class ServiceDescriptionProcessor extends AbstractInjectAnnotationProcess
                     serviceDescriptors,
                     originatingElements.toArray(io.micronaut.inject.ast.Element.EMPTY_ELEMENT_ARRAY)
             );
-
-            writeConfigurationMetadata();
         }
         return true;
     }
 
-    private void processContextConfigurerAnnotation(List<io.micronaut.inject.ast.Element> originatingElements, Element element, TypeElement typeElement) {
+    private void processContextConfigurerAnnotation(List<io.micronaut.inject.ast.Element> originatingElements,
+                                                    Element element,
+                                                    TypeElement typeElement) {
         AnnotationMetadata annotationMetadata = annotationUtils.getAnnotationMetadata(element);
         Optional<AnnotationValue<ContextConfigurer>> ann = annotationMetadata.findAnnotation(ContextConfigurer.class);
         if (ann.isPresent()) {
-            JavaClassElement javaClassElement = javaVisitorContext.getElementFactory().newClassElement(typeElement, annotationMetadata);
+            JavaClassElement javaClassElement = javaVisitorContext.getElementFactory()
+                    .newClassElement(typeElement, annotationMetadata);
             ContextConfigurerVisitor.assertNoConstructorForContextAnnotation(javaClassElement);
             List<? extends TypeMirror> interfaces = typeElement.getInterfaces();
             for (TypeMirror interfaceType : interfaces) {
@@ -122,7 +119,10 @@ public class ServiceDescriptionProcessor extends AbstractInjectAnnotationProcess
         AnnotationUtils.invalidateCache();
     }
 
-    private boolean processGeneratedAnnotation(List<io.micronaut.inject.ast.Element> originatingElements, Element element, TypeElement typeElement, String name) {
+    private boolean processGeneratedAnnotation(List<io.micronaut.inject.ast.Element> originatingElements,
+                                               Element element,
+                                               TypeElement typeElement,
+                                               String name) {
         Generated generated = element.getAnnotation(Generated.class);
         if (generated != null) {
             String serviceName = generated.service();
@@ -134,30 +134,5 @@ public class ServiceDescriptionProcessor extends AbstractInjectAnnotationProcess
             return true;
         }
         return false;
-    }
-
-    private void writeConfigurationMetadata() {
-        ConfigurationMetadataBuilder.getConfigurationMetadataBuilder().ifPresent(metadataBuilder -> {
-            try {
-                if (metadataBuilder.hasMetadata()) {
-                    ServiceLoader<ConfigurationMetadataWriter> writers = ServiceLoader.load(ConfigurationMetadataWriter.class, getClass().getClassLoader());
-
-                    try {
-                        for (ConfigurationMetadataWriter writer : writers) {
-                            try {
-                                writer.write(metadataBuilder, classWriterOutputVisitor);
-                            } catch (IOException e) {
-                                warning("Error occurred writing configuration metadata: %s", e.getMessage());
-                            }
-                        }
-                    } catch (ServiceConfigurationError e) {
-                        warning("Unable to load ConfigurationMetadataWriter due to : %s", e.getMessage());
-                    }
-                }
-            } finally {
-                ConfigurationMetadataBuilder.setConfigurationMetadataBuilder(null);
-            }
-        });
-
     }
 }
